Add 1V8 voltage regulator to fix internal UART voltage.
This commits adds a voltage regulator to change the voltage at the
VCCIO pin of the internal UART converter to 1V8.
This design was not tested in production/hardware yet but instead
is a hot fix to offer an alternative to the manufacturing of boards
with the wrong UART voltage.